fix having large video poster image
so the original poster image for the background video was about 600kb and it was by far the largest media on the website except the video itself. On the other hand, because video was the largest and also the first thing the user sees, it was important that at least the poster loads instantly, until the video is loaded. If both video and poster were heavy in size, we would have nothing except a button as the first view of the website. By using `webp` format, the size was reduced to around 60kb, which even on low network bandwidth such as a regular 3G, would be loaded in a second.