bump FontAwesome to version 6 (latest)
Firstly, updating to the latest version is a good idea, specially that the web version of font awesome was giving me codes for icons that didnt exist in local setup. Secondly, The original code used `all.css` approach which is not at all good for production, as it contains everything and makes the load time increadibliy slow. Now, we only included the base style and brands and only fonts and css files related to this are imported. Thirdly, all respective icons have been updated to match the version 6 naming.